@@ -44,13 +44,13 @@ class R {
 /// - fileExtname：.png
 class AssetResource {
   /// Creates an object to hold the asset resource’s metadata.
-  const AssetResource(this.assetName, {this.packageName}) : assert(assetName != null);
+  const AssetResource(this.assetName, {this.packageName});
 
   /// The name of the main asset from the set of asset resources to choose from.
   final String assetName;
 
   /// The name of the package from which the asset resource is included.
-  final String packageName;
+  final String? packageName;
 
   /// The name used to generate the key to obtain the asset resource. For local assets
   /// this is [assetName], and for assets from packages the [assetName] is
